+++ Circumference.hpp
@@ -11,7 +11,7 @@ private:
 	sf::Vector2f center;
 	float radius;
 public:
-	bool isInside(sf::Vector2f point);
+	bool isInside(sf::Vector2i point);
 
 	void draw(sf::RenderTarget* target, sf::Color color);
 };
